Split suit from it's tests according to the code review suggestion
